Simplifies and organizes code
