+++ PerformanceTests/PerformanceTests.csproj
@@ -330,12 +330,13 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BasicTypingTest.cs" />
-    <Compile Include="DiagnosticRunner.cs" />
+    <Compile Include="Tests\BasicTypingTest.cs" />
+    <Compile Include="Runners\DiagnosticRunner.cs" />
     <Compile Include="IDebuggableTest.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SimpleRunner.cs" />
+    <Compile Include="Runners\SimpleRunner.cs" />
+    <Compile Include="TestBase.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
